requirements.txt
bcrypt==4.0.1
blinker==1.6.2
cffi==1.15.1
click==8.1.6
cryptography==41.0.2
Flask==2.3.2
Flask-SQLAlchemy==3.0.5
greenlet==2.0.2
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
mysql==0.0.3
mysql-connector-python==8.1.0
mysqlclient==2.2.0
paramiko==3.3.1
protobuf==4.21.12
pycparser==2.21
PyNaCl==1.5.0
SQLAlchemy==2.0.19
sshtunnel==0.4.0
typing_extensions==4.7.1
Werkzeug==2.3.6
pymysql
